Add files via upload
"Built my first React project—a Login and Signup Form—and it was a great learning experience! This was my first time working with React, and I gained a lot from setting up components and managing state. I also used Vite for project setup, which made the development process faster and smoother. Overall, it was a valuable introduction to React!"